backend/requirements.txt

Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
kagglehub==0.1.7
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
